+++ crates/test-ui/ui/block_lifetimes_independent.rs
@@ -28,8 +28,7 @@ fn args_with_encoding<'a, 'b>(
 ) -> RcBlock<'static, fn(&'b i32, &'a i32)> {
     struct Enc<'a, 'b>(PhantomData<&'a i32>, PhantomData<&'b i32>);
     unsafe impl<'a, 'b> ManualBlockEncoding for Enc<'a, 'b> {
-        type Arguments = (&'a i32, &'b i32);
-        type Return = ();
+        type Signature = fn(&'a i32, &'b i32);
         const ENCODING_CSTR: &'static CStr = c"v24@?0^i8^i16";
     }
     RcBlock::with_encoding::<_, Enc<'a, 'b>>(f)
@@ -40,8 +39,7 @@ fn args_return_with_encoding<'a, 'b>(
 ) -> RcBlock<'static, fn(&'b i32) -> &'a i32> {
     struct Enc<'a, 'b>(PhantomData<&'a i32>, PhantomData<&'b i32>);
     unsafe impl<'a, 'b> ManualBlockEncoding for Enc<'a, 'b> {
-        type Arguments = (&'a i32,);
-        type Return = &'b i32;
+        type Signature = fn(&'a i32) -> &'b i32;
         const ENCODING_CSTR: &'static CStr = c"^i816@?0^i8";
     }
     RcBlock::with_encoding::<_, Enc<'a, 'b>>(f)
@@ -50,8 +48,7 @@ fn args_return_with_encoding<'a, 'b>(
 fn args_entire_with_encoding<'a, 'b>(f: impl Fn(&'a i32) + 'b) -> RcBlock<'a, fn(&'b i32)> {
     struct Enc<'a>(PhantomData<&'a i32>);
     unsafe impl<'a> ManualBlockEncoding for Enc<'a> {
-        type Arguments = (&'a i32,);
-        type Return = ();
+        type Signature = fn(&'a i32);
         const ENCODING_CSTR: &'static CStr = c"v16@?0^i8";
     }
     RcBlock::with_encoding::<_, Enc<'a>>(f)
@@ -62,8 +59,7 @@ fn return_entire_with_encoding<'a, 'b>(
 ) -> RcBlock<'a, fn() -> &'b i32> {
     struct Enc<'a>(PhantomData<&'a i32>);
     unsafe impl<'a> ManualBlockEncoding for Enc<'a> {
-        type Arguments = ();
-        type Return = &'a i32;
+        type Signature = fn() -> &'a i32;
         const ENCODING_CSTR: &'static CStr = c"^i8@?0";
     }
     RcBlock::with_encoding::<_, Enc<'a>>(f)